requirement.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
Django==5.2.3
dnspython==2.7.0
idna==3.10
mongoengine==0.29.1
numpy==2.2.6
opencv-python==4.12.0.88
paho-mqtt==2.1.0
pillow==11.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
pymongo==4.13.2
pyOpenSSL==25.1.0
pyzmq==27.0.0
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
